refactor: split quiz and public home frontend helpers
