+++ frontend/src/features/lms/pages/QuizGenPagePreview.tsx
@@ -1,0 +1,114 @@
+import type { CourseCard } from '@myway/shared';
+import { StatePanel } from '../components/StatePanel';
+import type { QuestionType } from './QuizGenPageSections';
+
+export function QuizGenPreviewPanel({
+  selectedCourse,
+  coverageSummary,
+  estimatedMinutes,
+  difficultyLabel,
+  questionCount,
+  activeTypes,
+}: {
+  selectedCourse: CourseCard | null;
+  coverageSummary: string;
+  estimatedMinutes: number;
+  difficultyLabel: string;
+  questionCount: number;
+  activeTypes: QuestionType[];
+}) {
+  return (
+    <section className="rounded-[30px] border border-slate-200 bg-white px-5 py-5 shadow-sm">
+      <div className="flex items-center justify-between gap-3">
+        <div>
+          <h3 className="text-[15px] font-bold text-slate-900">생성 미리보기</h3>
+          <p className="mt-1 text-[12px] text-slate-500">선택한 강의와 조건을 반영한 결과 구조를 먼저 보여줍니다.</p>
+        </div>
+        <div className="rounded-full bg-cyan-50 px-3 py-1 text-[11px] font-semibold text-cyan-700">{estimatedMinutes}분 예상</div>
+      </div>
+
+      {selectedCourse ? (
+        <div className="mt-4 grid gap-4 lg:grid-cols-[minmax(0,1fr)_280px]">
+          <div className="rounded-[26px] bg-[linear-gradient(135deg,#eff6ff_0%,#eef2ff_50%,#f5f3ff_100%)] px-5 py-5">
+            <div className="text-[12px] font-semibold text-cyan-700">선택 강의</div>
+            <div className="mt-1 text-[20px] font-extrabold tracking-[-0.03em] text-slate-900">{selectedCourse.title}</div>
+            <div className="mt-2 text-[13px] leading-6 text-slate-600">
+              {selectedCourse.category} · {selectedCourse.lecture_count}차시 · {selectedCourse.total_duration_minutes}분
+            </div>
+            <p className="mt-4 text-[13px] leading-7 text-slate-600">{coverageSummary}</p>
+            <div className="mt-5 flex flex-wrap gap-2">
+              <button type="button" className="rounded-full bg-cyan-600 px-4 py-2 text-[12px] font-semibold text-white transition hover:bg-cyan-500">
+                퀴즈 생성
+              </button>
+              <button type="button" className="rounded-full border border-slate-200 bg-white px-4 py-2 text-[12px] font-semibold text-slate-700 transition hover:border-cyan-200 hover:text-cyan-700">
+                저장하기
+              </button>
+            </div>
+          </div>
+
+          <div className="rounded-[26px] border border-slate-200 bg-slate-50 px-5 py-5">
+            <div className="text-[12px] font-semibold text-slate-400">설정 상태</div>
+            <div className="mt-3 space-y-3">
+              <div className="rounded-2xl bg-white px-4 py-3">
+                <div className="text-[11px] text-slate-400">난이도</div>
+                <div className="mt-1 text-[13px] font-semibold text-slate-900">{difficultyLabel}</div>
+              </div>
+              <div className="rounded-2xl bg-white px-4 py-3">
+                <div className="text-[11px] text-slate-400">문항 수</div>
+                <div className="mt-1 text-[13px] font-semibold text-slate-900">{questionCount}개</div>
+              </div>
+              <div className="rounded-2xl bg-white px-4 py-3">
+                <div className="text-[11px] text-slate-400">유형</div>
+                <div className="mt-1 flex flex-wrap gap-2">
+                  {activeTypes.length ? (
+                    activeTypes.map((type) => (
+                      <span key={type} className="rounded-full bg-cyan-50 px-2.5 py-1 text-[11px] font-semibold text-cyan-700">
+                        {type}
+                      </span>
+                    ))
+                  ) : (
+                    <span className="text-[12px] text-slate-500">유형을 하나 이상 선택하세요.</span>
+                  )}
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      ) : (
+        <div className="mt-4">
+          <StatePanel
+            compact
+            icon="ri-question-line"
+            tone="slate"
+            title="강의를 먼저 선택하세요."
+            description="강의를 고르면 문항 수, 난이도, 유형을 기준으로 생성 미리보기가 나타납니다."
+          />
+        </div>
+      )}
+    </section>
+  );
+}
+
+export function QuizGenRecommendationPanel() {
+  return (
+    <section className="rounded-[30px] border border-slate-200 bg-white px-5 py-5 shadow-sm">
+      <h3 className="flex items-center gap-2 text-[15px] font-bold text-slate-900">
+        <i className="ri-list-check-2 text-cyan-700" />
+        추천 구성
+      </h3>
+      <div className="mt-4 grid gap-3 md:grid-cols-2">
+        {[
+          { label: '핵심 개념 확인', value: '객관식 5문항' },
+          { label: '이해도 점검', value: '단답형 2문항' },
+          { label: '서술형 복습', value: '서술형 1문항' },
+          { label: '보조 복습', value: '빈칸 2문항' },
+        ].map((item) => (
+          <article key={item.label} className="rounded-[24px] border border-slate-200 bg-slate-50 px-4 py-4">
+            <div className="text-[11px] font-semibold text-slate-400">{item.label}</div>
+            <div className="mt-1 text-[14px] font-bold text-slate-900">{item.value}</div>
+          </article>
+        ))}
+      </div>
+    </section>
+  );
+}
