+++ frontend/src/lib/api-courses.ts
@@ -1,9 +1,7 @@
 import {
-  canEnroll,
   getCourseDetail,
   getDashboard,
   getLectureDetail,
-  getRoleLabel,
   type CourseCreateRequest,
   type CourseCard,
   type CourseDetail,
@@ -14,7 +12,9 @@ import {
   type NoticeCreateRequest,
 } from '@myway/shared';
 import { getFallbackUserId, getStoredAuth, request, unwrap } from './api-core';
+import { hydrateMissingLectureVideos, type EnrollmentItem } from './courses/course-network';
 import { mergeCourseDetailWithFallback, normalizeCourseId, normalizeLectureId } from './courses/course-mappers';
+export { canCurrentUserEnroll, getCurrentRoleLabel } from './courses/course-access';
 
 type CompleteLectureResponse = {
   lecture_id: string;
@@ -23,62 +23,6 @@ type CompleteLectureResponse = {
   completed_lectures: number;
   total_lectures: number;
 };
-
-type LectureVideoMappingResult = {
-  lecture_id: string;
-  asset_key?: string;
-  video_url?: string;
-};
-
-type EnrollmentItem = {
-  id: string;
-  user_id: string;
-  course_id: string;
-};
-
-async function hydrateMissingLectureVideos(detail: CourseDetail, token: string | null): Promise<CourseDetail> {
-  const missingLectureIds = detail.lectures
-    .filter((lecture) => !lecture.video_url && !lecture.video_asset_key)
-    .map((lecture) => lecture.id);
-
-  if (missingLectureIds.length === 0 || !token) {
-    return detail;
-  }
-
-  const mappingPairs = await Promise.all(
-    missingLectureIds.map(async (lectureId) => {
-      const response = await request<LectureVideoMappingResult>(
-        `/api/v1/media/lecture-video/${encodeURIComponent(lectureId)}`,
-        undefined,
-        token,
-      );
-      if (!response?.success || !response.data) {
-        return null;
-      }
-      return [lectureId, response.data] as const;
-    }),
-  );
-
-  const mappingMap = new Map(mappingPairs.filter(Boolean) as ReadonlyArray<readonly [string, LectureVideoMappingResult]>);
-  if (mappingMap.size === 0) {
-    return detail;
-  }
-
-  return {
-    ...detail,
-    lectures: detail.lectures.map((lecture) => {
-      const mapping = mappingMap.get(lecture.id);
-      if (!mapping) {
-        return lecture;
-      }
-      return {
-        ...lecture,
-        video_url: lecture.video_url ?? mapping.video_url,
-        video_asset_key: lecture.video_asset_key ?? mapping.asset_key,
-      };
-    }),
-  };
-}
 
 export async function loadCourses(sessionToken?: string | null): Promise<CourseCard[]> {
   const token = sessionToken ?? getStoredAuth()?.session_token ?? null;
@@ -279,12 +223,3 @@ export async function enrollCourse(
   return response?.success && response.data ? response.data : null;
 }
 
-export function canCurrentUserEnroll(): boolean {
-  const storedAuth = getStoredAuth();
-  return storedAuth ? canEnroll(storedAuth.user.role) : false;
-}
-
-export function getCurrentRoleLabel(): string {
-  const storedAuth = getStoredAuth();
-  return storedAuth ? getRoleLabel(storedAuth.user.role) : '게스트';
-}
